Replace layouts with embedded counterparts
