+++ src/layouts/default.vue
@@ -1,19 +1,28 @@
 <template>
   <div class="app">
-    <HeaderSection :is-embedded="false" />
-    <main>
+    <MigrationNotice v-show="isReferredFromCc" />
+    <HeaderSection />
+    <main class="embedded">
       <Nuxt />
     </main>
-    <FooterSection />
   </div>
 </template>
 <script>
-const defaultPage = {
-  name: 'default',
-  layout: 'default',
+import { mapState } from 'vuex'
+
+import iframeHeight from '~/mixins/iframe-height'
+import i18nSync from '~/mixins/i18n-sync'
+
+import { NAV } from '~/constants/store-modules'
+
+const embeddedPage = {
+  name: 'embedded',
+  layout: 'embedded',
+  mixins: [iframeHeight, i18nSync],
   head() {
     return this.$nuxtI18nHead({ addSeoAttributes: true, addDirAttribute: true })
   },
+  computed: { ...mapState(NAV, ['isReferredFromCc']) },
 }
-export default defaultPage
+export default embeddedPage
 </script>
